feat(config): Externalize and validate JWT configuration
- Makes JWT secret key and expiration time configurable using @Value
  and application.properties (app.jwt.*).
- Uses Lombok @RequiredArgsConstructor for cleaner injection in JwtTokenProvider.
- Adds @PostConstruct validation in JwtTokenProvider to ensure the
  secret is present

refactor(security): Consolidate security configuration beans

- Moves AuthenticationManager, DaoAuthenticationProvider, and
  PasswordEncoder bean definitions from separate config classes
  into the main SecurityConfig class for better cohesion.
- Deletes AuthenticationManagerConfig and PasswordEncoderConfig.
- Updates CSRF configuration syntax to use method reference.
