+++ src/main/java/com/ftgo/user/userservice/security/config/JwtTokenProvider.java
@@ -5,29 +5,62 @@ import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
 import io.jsonwebtoken.io.Decoders;
 import io.jsonwebtoken.security.Keys;
+import jakarta.annotation.PostConstruct;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Component;
+import org.springframework.util.StringUtils;
 
 import javax.crypto.SecretKey;
 import java.util.Date;
 import java.util.List;
 
-
+@Slf4j
+@RequiredArgsConstructor
 @Component
 public class JwtTokenProvider {
 
 
-    // In a production environment, move this secret to a secure external configuration.
-    private static final String jwtSecret = "fBsWwVrv0BNVpGIRYAjCeXn8nPCp5nHNfMvQY+PI5H9NE37H6ForsPalQRZi01d7zmfLVxe35UIE1qxwWVZxLw==";
-    private static final long jwtExpirationInMs = 3600000; // 1 hour
-    private final SecretKey key;
+    @Value("${app.jwt.secret}")
+    private String jwtSecret;
+
+    @Value("${app.jwt.expiration-ms}")
+    private long jwtExpirationInMs;
+    private SecretKey key;
 
 
-    public JwtTokenProvider() {
-        byte[] keyBytes = Decoders.BASE64.decode(jwtSecret);
+    /**
+     * Initializes the SecretKey after properties are injected.
+     * Validates that the secret key is provided and sufficiently strong.
+     */
+    @PostConstruct
+    public void init() {
+        if (!StringUtils.hasText(jwtSecret)) {
+            log.error("FATAL: JWT secret key is not configured in properties (app.jwt.secret). Application cannot start securely.");
+            throw new IllegalStateException("JWT secret key ('app.jwt.secret') cannot be empty.");
+        }
+
+        byte[] keyBytes;
+        try {
+            keyBytes = Decoders.BASE64.decode(jwtSecret);
+        } catch (IllegalArgumentException e) {
+            log.error("FATAL: JWT secret key ('app.jwt.secret') is not valid Base64: {}", e.getMessage());
+            throw new IllegalStateException("JWT secret key ('app.jwt.secret') must be a valid Base64 encoded string.", e);
+        }
+
         this.key = Keys.hmacShaKeyFor(keyBytes);
+        log.info("JwtTokenProvider initialized successfully.");
     }
+
+    /**
+     * Generates a JWT token for the given UserDetails.
+     *
+     * @param userDetails The user details object containing username and authorities.
+     * @return A JWT token string.
+     */
     public String generateToken(UserDetails userDetails) {
         Date now = new Date();
         Date expiryDate = new Date(now.getTime() + jwtExpirationInMs);
