+++ src/main/java/com/ftgo/user/userservice/security/config/SecurityConfig.java
@@ -1,10 +1,17 @@
 package com.ftgo.user.userservice.security.config;
 
+import com.ftgo.user.userservice.security.service.CustomUserDetailsService;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
 import org.springframework.security.config.Customizer;
+import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
 import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 
@@ -16,13 +23,46 @@ public class SecurityConfig {
     @Bean
     public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
         http
-                .csrf(csrf -> csrf.disable()) // For stateless APIs, disable CSRF protection
+                .csrf(AbstractHttpConfigurer::disable)
                 .authorizeHttpRequests(auth -> auth
                         .requestMatchers("/auth/login", "/auth/register").permitAll() // Public endpoints
                         .anyRequest().authenticated() // Secure other endpoints
                 )
                 .httpBasic(Customizer.withDefaults()); // Use basic authentication as a fallback for testing
         return http.build();
+    }
+
+    /**
+     * Configures a `DaoAuthenticationProvider` bean.
+     *
+     * @param userDetailsService The custom user details service used to load user-specific data.
+     * @param passwordEncoder The password encoder used to encode and verify passwords.
+     * @return A configured `DaoAuthenticationProvider` instance.
+     */
+    @Bean
+    public DaoAuthenticationProvider authenticationProvider(CustomUserDetailsService userDetailsService,
+                                                            PasswordEncoder passwordEncoder) {
+        DaoAuthenticationProvider provider = new DaoAuthenticationProvider();
+        provider.setUserDetailsService(userDetailsService);
+        provider.setPasswordEncoder(passwordEncoder);
+        return provider;
+    }
+
+    /**
+     * Configures the `AuthenticationManager` bean.
+     *
+     * @param config The authentication configuration used to create the authentication manager.
+     * @return A configured `AuthenticationManager` instance.
+     * @throws Exception If an error occurs while creating the authentication manager.
+     */
+    @Bean
+    public AuthenticationManager authenticationManager(AuthenticationConfiguration config) throws Exception {
+        return config.getAuthenticationManager();
+    }
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
     }
 }
