+++ src/main/resources/application.properties
@@ -11,3 +11,13 @@ spring.datasource.password=hamidreza80
 spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
+
+# ===============================================
+# Application JWT Configuration
+# ===============================================
+
+# JWT secret key (should be kept secret and secure)
+app.jwt.secret=fBsWwVrv0BNVpGIRYAjCeXn8nPCp5nHNfMvQY+PI5H9NE37H6ForsPalQRZi01d7zmfLVxe35UIE1qxwWVZxLw==
+
+# Token expiration time in milliseconds (e.g., 3600000 = 1 hour, 86400000 = 1 day)
+app.jwt.expiration-ms=3600000
